requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
ItsDangerous==2.1.2
click==8.1.7
Pillow==9.3.0
python_version >= "3.9"
gunicorn==21.2.0  # Optional, if deploying on a server
